Rough packaging and some classes for expansion for unique abilities for various characters
Added randomization of names for Monsters

Added name prompt for player

Made MonsterCompendium a class

Made a DungeonMap class to contain the DungeonRooms

Plenty still to do/not done, suggest leaving this branch as a branch or forking in stead of merging at this stage.
