+++ src/toons/Combat.java
@@ -1,0 +1,7 @@
+package toons;
+
+public interface Combat
+{
+   public boolean monsterAttack(Player player);
+
+}
